#132 Enable posting of outlook without journalEntries
Mach the product spec, by enabling outlooks as well as journal entries be posted and storred independantly inside a reflection. If only a journal entry is posted, there will be no calls to the database to save an outlook as an empty array. If only outlook is posted, there will be no calls to the database to save an empty string as a journal entry .

Rename selectedOptions to selectedOptionIds to reflect what it holds. Declare selectedOptionIds data type as an array of options from the start to match backend expectations and avoid data type reassigning.

use Number() instead of `+` to cast the correct type to an option id for better code readability and according to Airbnb's style guide  https://github.com/airbnb/javascript#coercion--numbers

Resolves #132
Resolves #135