#347: All fields now editable
This version now shows an editing form that allows for users to edit every field of their profile. Instead of an "add social network"/"remove social network" control, if a user blanks out the "user name" field for a social network, that removes it from the list of social networks. The control that allows a user to change public visibility only shows if there's something in the "user name" field. And there's a "cancel editing" button that restores any changes to the user profile to the version that comes from the API.