#415 Define test vars, rearrange function order
Rearrange the function order in the API for both `programsService.ts` and its test file. Also, add in variables that will need to be defined for each of the tests in that test file. Finally, add in the relevant TypeScript types to better guide the structure of I/O data.