#415 Time zone support
Modify programsService and the tests for programsService and programsRouter to support specifying time zones for programs. This makes it so all the math works out. All relevant program activity dates will be sent back as UTC date strings, making it easier on the frontend.

However, it will require the addition of an external library, [luxon](https://moment.github.io/luxon), which is the successor to the very popular [moment.js](https://momentjs.com).