#415 Return date strings as strings from mySQL
By default, knex will use the default behavior of mysqljs and cast MySQL DATE and DATETIME values as JavaScript Date()s. This makes them troublesome to parse and causes issues with the differences between application behavior in production vs in test.

Disable this behavior by telling the mysql client to return them as strings. While we’re at it, specify the time zone for the connection to the SQL server, since we’re already incorporating the `luxon` library.