#415 Mark optional fields in models optional
For the fields that are optional that don’t need to be retrieved from the database, mark them as optional in the models file.

From here, we will need to update the database calls to retrieve only the necessary, non-optional fields, then we need to update the tests to remove the optional fields as well.