#437 Modifications to curriculum activities
I wrote some code to generate the SQL to insert for the example curriculum activities to ensure every activity is placed on the correct day.

Update all the activities to better align their names, descriptions, and times with the current cohort’s activities, and add any missing activities.

Add an additional curriculum value (“Custom Curriculum”) to allow for unstructured curricula.

Add an additional program value (“Cohort 5”) with some non-final dates so we have more than one program with which we test the front-end features.

Update the list of program activities to include more types of activities.

Improve spacing and layout so SQL is easier to debug if necessary.