#428 Transform Program into a class
The return value of Program() and the name didn’t match up before; `Program` didn’t return a Program, it returned a calendar.

Switching Program to a class allows us to use it both as a data type and as a way to show the calendar, thereby supporting being able to switch programs and show a different calendar.

(However, if we want to make this mutable... that’ll be more work down the road.)